widgets/charts: implement base polar chart class
