+++ Software/PC_Application/Traces/smithchartdialog.ui
@@ -117,6 +117,16 @@
           <item row="1" column="1">
            <widget class="SIUnitEdit" name="zoomReflection"/>
           </item>
+          <item row="2" column="0">
+           <widget class="QLabel" name="label_6">
+            <property name="text">
+             <string>Offset real axis</string>
+            </property>
+           </widget>
+          </item>
+          <item row="2" column="1">
+           <widget class="SIUnitEdit" name="offsetRealAxis"/>
+          </item>
          </layout>
         </widget>
        </item>
